Folders reorganization. Remove CounterView
